HW7/requirements.txt
flask
cassandra-driver
pyspark==3.4.3
redis